fix: fix and move tests
