+++ apps/meteor/app/federation-v2/server/domain/FederatedRoom.ts
@@ -11,9 +11,9 @@ export class FederatedRoom {
 	public internalReference: IRoom;
 
 	// eslint-disable-next-line
-	private constructor() {} 
+	protected constructor() {} 
 
-	private static generateTemporaryName(normalizedExternalId: string): string {
+	protected static generateTemporaryName(normalizedExternalId: string): string {
 		return `Federation-${normalizedExternalId}`;
 	}
 
@@ -23,7 +23,7 @@ export class FederatedRoom {
 		creator: FederatedUser,
 		type: RoomType,
 		name?: string,
-		members?: IUser[],
+		members?: FederatedUser[],
 	): FederatedRoom {
 		const roomName = name || FederatedRoom.generateTemporaryName(normalizedExternalId);
 		return Object.assign(new FederatedRoom(), {
@@ -42,34 +42,8 @@ export class FederatedRoom {
 		return new FederatedRoom();
 	}
 
-	public isDirectMessage(): boolean {
-		return this.internalReference?.t === RoomType.DIRECT_MESSAGE;
-	}
-
-	public setRoomType(type: RoomType): void {
-		if (this.isDirectMessage()) {
-			throw new Error('Its not possible to change a direct message type');
-		}
-		this.internalReference.t = type;
-	}
-
-	public changeRoomName(name: string): void {
-		if (this.isDirectMessage()) {
-			throw new Error('Its not possible to change a direct message name');
-		}
-		this.internalReference.name = name;
-		this.internalReference.fname = name;
-	}
-
-	public changeRoomTopic(topic: string): void {
-		if (this.isDirectMessage()) {
-			throw new Error('Its not possible to change a direct message topic');
-		}
-		this.internalReference.description = topic;
-	}
-
 	public getMembers(): IUser[] {
-		return this.isDirectMessage() && this.members && this.members.length > 0 ? this.members.map((user) => user.internalReference) : [];
+		return this.members && this.members.length > 0 ? this.members.map((user) => user.internalReference) : [];
 	}
 
 	public isFederated(): boolean {
