+++ apps/meteor/app/federation-v2/server/domain/FederatedUser.ts
@@ -14,7 +14,7 @@ export class FederatedUser {
 	public existsOnlyOnProxyServer: boolean;
 
 	// eslint-disable-next-line
-	private constructor() {}
+	protected constructor() {}
 
 	public static createInstance(externalId: string, params: IFederatedUserCreationParams): FederatedUser {
 		return Object.assign(new FederatedUser(), {
