+++ apps/meteor/app/federation-v2/server/infrastructure/matrix/definitions/MatrixEventType.ts
@@ -2,4 +2,6 @@ export enum MatrixEventType {
 	ROOM_CREATED = 'm.room.create',
 	ROOM_MEMBERSHIP_CHANGED = 'm.room.member',
 	ROOM_MESSAGE_SENT = 'm.room.message',
+	ROOM_JOIN_RULES_CHANGED = 'm.room.join_rules',
+	ROOM_NAME_CHANGED = 'm.room.name',
 }
