+++ apps/meteor/client/importPackages.ts
@@ -43,7 +43,6 @@ import '../app/apps/client';
 import '../app/slackbridge/client';
 import '../app/slashcommands-archiveroom/client';
 import '../app/slashcommand-asciiarts/client';
-import '../app/slashcommands-federation/client';
 import '../app/slashcommands-create/client';
 import '../app/slashcommands-hide/client';
 import '../app/slashcommands-invite/client';
