+++ apps/meteor/ee/app/federation-v2/server/infrastructure/rocket-chat/adapters/Room.ts
@@ -1,10 +1,33 @@
+import { IRoom } from '@rocket.chat/core-typings';
+
 import { FederatedRoom } from '../../../../../../../app/federation-v2/server/domain/FederatedRoom';
 import { RocketChatRoomAdapter } from '../../../../../../../app/federation-v2/server/infrastructure/rocket-chat/adapters/Room';
 import { MatrixBridgedRoom } from '../../../../../../../app/models/server';
 import { Rooms, Subscriptions } from '../../../../../../../app/models/server/raw';
+import { FederatedRoomEE } from '../../../domain/FederatedRoom';
 
 export class RocketChatRoomAdapterEE extends RocketChatRoomAdapter {
-	public async updateRoomType(federatedRoom: FederatedRoom): Promise<void> {
+	public async getFederatedRoomByExternalId(externalRoomId: string): Promise<FederatedRoomEE | undefined> {
+		const internalBridgedRoomId = MatrixBridgedRoom.getId(externalRoomId);
+		if (!internalBridgedRoomId) {
+			return;
+		}
+		const room = await Rooms.findOneById(internalBridgedRoomId);
+
+		return this.createFederatedRoomEEInstance(externalRoomId, room);
+	}
+
+	public async getFederatedRoomByInternalId(internalRoomId: string): Promise<FederatedRoomEE | undefined> {
+		const externalRoomId = MatrixBridgedRoom.getMatrixId(internalRoomId);
+		if (!externalRoomId) {
+			return;
+		}
+		const room = await Rooms.findOneById(internalRoomId);
+
+		return this.createFederatedRoomEEInstance(externalRoomId, room);
+	}
+
+	public async updateRoomType(federatedRoom: FederatedRoomEE): Promise<void> {
 		await Rooms.update({ _id: federatedRoom.internalReference._id }, { $set: { t: federatedRoom.internalReference.t } });
 		await Subscriptions.update(
 			{ rid: federatedRoom.internalReference._id },
@@ -25,15 +48,23 @@ export class RocketChatRoomAdapterEE extends RocketChatRoomAdapter {
 		);
 	}
 
-	public async updateRoomTopic(federatedRoom: FederatedRoom): Promise<void> {
+	public async updateRoomTopic(federatedRoom: FederatedRoomEE): Promise<void> {
 		await Rooms.update(
 			{ _id: federatedRoom.internalReference._id },
 			{ $set: { description: federatedRoom.internalReference.description } },
 		);
 	}
 
-	public async updateFederatedRoomByInternalRoomId(internalRoomId: string, federatedRoom: FederatedRoom): Promise<void> {
+	public async updateFederatedRoomByInternalRoomId(internalRoomId: string, federatedRoom: FederatedRoomEE): Promise<void> {
 		MatrixBridgedRoom.upsert({ rid: internalRoomId }, { rid: internalRoomId, mri: federatedRoom.externalId });
 		await Rooms.setAsFederated(internalRoomId);
+	}
+
+	private createFederatedRoomEEInstance(externalRoomId: string, room: IRoom): FederatedRoomEE {
+		const federatedRoom = FederatedRoomEE.build();
+		federatedRoom.externalId = externalRoomId;
+		federatedRoom.internalReference = room;
+
+		return federatedRoom;
 	}
 }
