+++ apps/meteor/ee/app/federation-v2/server/infrastructure/rocket-chat/converters/RoomSender.ts
@@ -1,0 +1,20 @@
+import { FederationRoomInviteUserDto } from '../../../application/input/RoomSenderDto';
+
+export class FederationRoomSenderConverterEE {
+	public static toRoomInviteUserDto(
+		internalInviterId: string,
+		internalRoomId: string,
+		externalInviteeId: string,
+	): FederationRoomInviteUserDto {
+		const normalizedInviteeId = externalInviteeId.replace('@', '');
+		const inviteeUsernameOnly = externalInviteeId.split(':')[0]?.replace('@', '');
+
+		return Object.assign(new FederationRoomInviteUserDto(), {
+			internalInviterId,
+			internalRoomId,
+			rawInviteeId: externalInviteeId,
+			normalizedInviteeId,
+			inviteeUsernameOnly,
+		});
+	}
+}
