+++ apps/meteor/ee/app/federation-v2/server/infrastructure/rocket-chat/slash-commands/index.ts
@@ -6,8 +6,6 @@ import {
 
 const EE_FEDERATION_COMMANDS = {
 	...FEDERATION_COMMANDS,
-	// 'setup-room': async () => {},
-	// 'invite': async () => {},
 };
 
 function federation(providedCommand: string, stringParams: string | undefined, item: Record<string, any>): void {
