+++ apps/meteor/tests/unit/app/federation-v2/unit/application/RoomServiceReceiver.spec.ts
@@ -16,9 +16,6 @@ describe('Federation - Application - FederationRoomServiceReceiver', () => {
 		createFederatedRoom: sinon.stub(),
 		removeUserFromRoom: sinon.stub(),
 		addUserToRoom: sinon.stub(),
-		updateRoomType: sinon.stub(),
-		updateRoomName: sinon.stub(),
-		updateRoomTopic: sinon.stub(),
 	};
 	const userAdapter = {
 		getFederatedUserByExternalId: sinon.stub(),
@@ -51,9 +48,6 @@ describe('Federation - Application - FederationRoomServiceReceiver', () => {
 		roomAdapter.createFederatedRoom.reset();
 		roomAdapter.removeUserFromRoom.reset();
 		roomAdapter.addUserToRoom.reset();
-		roomAdapter.updateRoomType.reset();
-		roomAdapter.updateRoomName.reset();
-		roomAdapter.updateRoomTopic.reset();
 		userAdapter.getFederatedUserByExternalId.reset();
 		userAdapter.createFederatedUser.reset();
 		messageAdapter.sendMessage.reset();
@@ -231,7 +225,10 @@ describe('Federation - Application - FederationRoomServiceReceiver', () => {
 				normalizedInviteeId: 'normalizedInviteeId',
 			} as any);
 
-			const room = FederatedRoom.createInstance('externalRoomId', 'normalizedRoomId', inviter as FederatedUser, RoomType.CHANNEL);
+			const room = FederatedRoom.createInstance('externalRoomId', 'normalizedRoomId', inviter as FederatedUser, RoomType.CHANNEL, '', [
+				inviter,
+				invitee,
+			] as FederatedUser[]);
 			expect(roomAdapter.createFederatedRoom.calledWith(room)).to.be.true;
 			expect(bridge.joinRoom.calledWith('externalRoomId', 'externalInviteeId')).to.be.true;
 		});
@@ -325,112 +322,6 @@ describe('Federation - Application - FederationRoomServiceReceiver', () => {
 			} as any);
 
 			expect(messageAdapter.sendMessage.calledWith({}, 'text', {})).to.be.true;
-		});
-	});
-
-	describe('#changeJoinRules()', () => {
-		it('should NOT change the room type if the room does not exists', async () => {
-			roomAdapter.getFederatedRoomByExternalId.resolves(undefined);
-			await service.changeJoinRules({
-				roomType: RoomType.CHANNEL,
-			} as any);
-
-			expect(roomAdapter.updateRoomType.called).to.be.false;
-		});
-
-		it('should NOT change the room type if it exists and is a direct message', async () => {
-			const room = FederatedRoom.build();
-			room.internalReference = {} as any;
-			room.internalReference.t = RoomType.DIRECT_MESSAGE;
-			roomAdapter.getFederatedRoomByExternalId.resolves(room);
-			await service.changeJoinRules({
-				roomType: RoomType.CHANNEL,
-			} as any);
-
-			expect(roomAdapter.updateRoomType.called).to.be.false;
-		});
-
-		it('should change the room type if it exists and is NOT a direct message', async () => {
-			const room = FederatedRoom.build();
-			room.internalReference = {} as any;
-			room.internalReference.t = RoomType.PRIVATE_GROUP;
-			roomAdapter.getFederatedRoomByExternalId.resolves(room);
-			await service.changeJoinRules({
-				roomType: RoomType.CHANNEL,
-			} as any);
-			room.internalReference.t = RoomType.CHANNEL;
-			expect(roomAdapter.updateRoomType.calledWith(room)).to.be.true;
-		});
-	});
-
-	describe('#changeRoomName()', () => {
-		it('should NOT change the room name if the room does not exists', async () => {
-			roomAdapter.getFederatedRoomByExternalId.resolves(undefined);
-			await service.changeRoomName({
-				normalizedRoomName: 'normalizedRoomName',
-			} as any);
-
-			expect(roomAdapter.updateRoomName.called).to.be.false;
-		});
-
-		it('should NOT change the room name if it exists and is a direct message', async () => {
-			const room = FederatedRoom.build();
-			room.internalReference = {} as any;
-			room.internalReference.t = RoomType.DIRECT_MESSAGE;
-			roomAdapter.getFederatedRoomByExternalId.resolves(room);
-			await service.changeRoomName({
-				normalizedRoomName: 'normalizedRoomName',
-			} as any);
-
-			expect(roomAdapter.updateRoomName.called).to.be.false;
-		});
-
-		it('should change the room name if it exists and is NOT a direct message', async () => {
-			const room = FederatedRoom.build();
-			room.internalReference = {} as any;
-			room.internalReference.t = RoomType.PRIVATE_GROUP;
-			roomAdapter.getFederatedRoomByExternalId.resolves(room);
-			await service.changeRoomName({
-				roomnormalizedRoomNameType: 'normalizedRoomName',
-			} as any);
-			room.internalReference.name = 'normalizedRoomName';
-			room.internalReference.fname = 'normalizedRoomName';
-			expect(roomAdapter.updateRoomName.calledWith(room)).to.be.true;
-		});
-	});
-
-	describe('#changeRoomTopic()', () => {
-		it('should NOT change the room topic if the room does not exists', async () => {
-			roomAdapter.getFederatedRoomByExternalId.resolves(undefined);
-			await service.changeRoomTopic({
-				roomTopic: 'roomTopic',
-			} as any);
-
-			expect(roomAdapter.updateRoomTopic.called).to.be.false;
-		});
-
-		it('should NOT change the room topic if it exists and is a direct message', async () => {
-			const room = FederatedRoom.build();
-			room.internalReference = {} as any;
-			room.internalReference.t = RoomType.DIRECT_MESSAGE;
-			roomAdapter.getFederatedRoomByExternalId.resolves(room);
-			await service.changeRoomTopic({
-				roomTopic: 'roomTopic',
-			} as any);
-
-			expect(roomAdapter.updateRoomTopic.called).to.be.false;
-		});
-
-		it('should change the room topic if it exists and is NOT a direct message', async () => {
-			const room = FederatedRoom.build();
-			room.internalReference = {} as any;
-			room.internalReference.t = RoomType.PRIVATE_GROUP;
-			roomAdapter.getFederatedRoomByExternalId.resolves(room);
-			await service.changeRoomTopic({
-				roomTopic: 'roomTopic',
-			} as any);
-			room.internalReference.description = 'roomTopic';
-			expect(roomAdapter.updateRoomTopic.calledWith(room)).to.be.true;
 		});
 	});
 });
